move função formatFieldName de src/components para src/helpers
